Move metric sampling logic to `BalacnerHandler`
